@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Get the Data from csv-file to generate to output."""
 # -*- coding: utf-8 -*-
 #
@@ -70,13 +69,13 @@ class GetData:
         """Get the data from the csv-file."""
         # global orders_file
         try:
-            with open(self.file_name, "r") as orders_file:
+            with open(self.file_name) as orders_file:
                 orders = csv.reader(orders_file, delimiter=";", quotechar='"')
                 data = list(orders)
 
                 return data
 
-        except IOError:
+        except OSError:
             logger.debug(
                 "The File with the data '{}' does not exists".format(
                     self.file_name
