@@ -1,27 +1,54 @@
-# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2019-2020 Jürgen Mülbert. All rights reserved.
+#
+# Licensed under the EUPL, Version 1.2 or – as soon they
+# will be approved by the European Commission - subsequent
+# versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the
+# Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/page/eupl-text-11-12
+#
+# Unless required by applicable law or agreed to in
+# writing, software distributed under the Licence is
+# distributed on an "AS IS" basis,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+# express or implied.
+# See the Licence for the specific language governing
+# permissions and limitations under the Licence.
+#
+# Lizenziert unter der EUPL, Version 1.2 oder - sobald
+#  diese von der Europäischen Kommission genehmigt wurden -
+# Folgeversionen der EUPL ("Lizenz");
+# Sie dürfen dieses Werk ausschließlich gemäß
+# dieser Lizenz nutzen.
+# Eine Kopie der Lizenz finden Sie hier:
+#
+# https://joinup.ec.europa.eu/page/eupl-text-11-12
+#
+# Sofern nicht durch anwendbare Rechtsvorschriften
+# gefordert oder in schriftlicher Form vereinbart, wird
+# die unter der Lizenz verbreitete Software "so wie sie
+# ist", OHNE JEGLICHE GEWÄHRLEISTUNG ODER BEDINGUNGEN -
+# ausdrücklich oder stillschweigend - verbreitet.
+# Die sprachspezifischen Genehmigungen und Beschränkungen
+# unter der Lizenz sind dem Lizenztext zu entnehmen.
+#
 import os
-
-from jmopenorders.core.logger import logger
 
 from . import cleanoutputdir
 from . import generateorders
 from . import getdata
 from . import getserviceperson
+from jmopenorders.core.logger import logger
 
 
-<<<<<<< Updated upstream:jmopenorders/api/report.py
-def Report(
-        personfile="names.csv",
-        datafile="data.csv",
-        inputpath="home",
-        outputpath="out",
-=======
 def report(
     personfile: str = "names.csv",
     datafile: str = "data.csv",
     inputpath: str = "home",
     outputpath: str = "out",
->>>>>>> Stashed changes:src/jmopenorders/api/report.py
 ) -> str:
 
     logger.debug("executing report command")
