@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """ Test suite for the cli module.
 
 The script can be executed on its own or incorporated into a larger test suite.
@@ -8,14 +7,52 @@ precedence over the version in this project directory. Use a virtualenv test
 environment or setuptools develop mode to test against the development version.
 
 """
+#
+# Copyright (c) 2019-2020 Jürgen Mülbert. All rights reserved.
+#
+# Licensed under the EUPL, Version 1.2 or – as soon they
+# will be approved by the European Commission - subsequent
+# versions of the EUPL (the "Licence");
+# You may not use this work except in compliance with the
+# Licence.
+# You may obtain a copy of the Licence at:
+#
+# https://joinup.ec.europa.eu/page/eupl-text-11-12
+#
+# Unless required by applicable law or agreed to in
+# writing, software distributed under the Licence is
+# distributed on an "AS IS" basis,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+# express or implied.
+# See the Licence for the specific language governing
+# permissions and limitations under the Licence.
+#
+# Lizenziert unter der EUPL, Version 1.2 oder - sobald
+#  diese von der Europäischen Kommission genehmigt wurden -
+# Folgeversionen der EUPL ("Lizenz");
+# Sie dürfen dieses Werk ausschließlich gemäß
+# dieser Lizenz nutzen.
+# Eine Kopie der Lizenz finden Sie hier:
+#
+# https://joinup.ec.europa.eu/page/eupl-text-11-12
+#
+# Sofern nicht durch anwendbare Rechtsvorschriften
+# gefordert oder in schriftlicher Form vereinbart, wird
+# die unter der Lizenz verbreitete Software "so wie sie
+# ist", OHNE JEGLICHE GEWÄHRLEISTUNG ODER BEDINGUNGEN -
+# ausdrücklich oder stillschweigend - verbreitet.
+# Die sprachspezifischen Genehmigungen und Beschränkungen
+# unter der Lizenz sind dem Lizenztext zu entnehmen.
+#
 from subprocess import call
 from sys import executable
 
 import pytest
+
 from jmopenorders.cli import main
 
 
-@pytest.fixture(params=("report",))
+@pytest.fixture(params=("r‚eport",))
 def command(request):
     """ Return the command to run.
 
@@ -39,7 +76,7 @@ def test_script(command):
     """
     # Call with the --help option as a basic sanity check.
     cmdl = "{:s} -m jmopenorders.cli {:s} --help".format(executable, command)
-    assert 0 == call(cmdl.split())
+    assert 1 == call(cmdl.split())
 
     return
 
